refactor: set custom DNS server only once (#821)
* refactor: set custom DNS server only once

Every time an HTTP client is created, an attempt is made to modify the `dialer.Resolver`.

So now it only attempts to change the `dialer.Resolver` once when necessary (e.g. by passing the `-dns` parameter or running the `WaitForNetworkConnected` function).

Also, rename `dns.go` and `dns_test.go` to `net_resolver.go` and `net_resolver_test.go`

---

Move `wait_net.go` to the `internal` package to prevent it from being mistaken as a DNS provider.

Ignore false positive.

* test(net_resolver): clean up after testing
