+++ util/http_client_util.go
@@ -31,7 +31,6 @@ var defaultTransport = &http.Transport{
 
 // CreateHTTPClient Create Default HTTP Client
 func CreateHTTPClient() *http.Client {
-	dialer.Resolver = CustomDNSResolver()
 	// SkipVerfiry
 	defaultTransport.TLSClientConfig = &tls.Config{InsecureSkipVerify: os.Getenv(SkipVerifyENV) == "true"}
 	return &http.Client{
@@ -74,7 +73,6 @@ var noProxyTcp6Transport = &http.Transport{
 
 // CreateNoProxyHTTPClient Create NoProxy HTTP Client
 func CreateNoProxyHTTPClient(network string) *http.Client {
-	dialer.Resolver = CustomDNSResolver()
 	if network == "tcp6" {
 		// SkipVerfiry
 		noProxyTcp6Transport.TLSClientConfig = &tls.Config{InsecureSkipVerify: os.Getenv(SkipVerifyENV) == "true"}
